Day 14 part 1
Too inefficient for Part 2. I wasted a bunch of time on the linked
list thinking the issue was the repeated copying from inserts, but in
fact it's just the length of the string: I convinced myself the exponent
was <1.5 (because each *two* chars produce *one* char), where in fact
it's <2 (because each pair produces one, yes, but each char appears in
two adjacent pairs). That grows way to big to represent the full list.
Luckily we don't need to.